requirements.txt
groq
python-dotenv
pydantic
numpy
pytest
colorama
